requirements.txt
pandas
numpy 
seaborn
requests
-e .